Move the database fixture in a common directory.
